Index file renames Part 2
