+++ Composer/packages/client/__tests__/utils/fileUtil.test.js
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { getExtension, getBaseName, upperCaseName } from '../../src/utils';
+import { getExtension, getBaseName, upperCaseName } from '../../src/utils/fileUtil';
 
 const files = ['a.text', 'a.b.text', 1];
