+++ Composer/packages/client/src/components/TestController/TestController.tsx
@@ -14,7 +14,7 @@ import { BotStatus, LuisConfig } from '../../constants';
 import { isAbsHosted } from '../../utils/envUtil';
 import { getReferredFiles } from '../../utils/luUtil';
 import useNotifications from '../../pages/notifications/useNotifications';
-import { navigateTo, openInEmulator } from '../../utils';
+import { navigateTo, openInEmulator } from '../../utils/navigation';
 
 import { PublishLuisDialog } from './publishDialog';
 import { bot, botButton } from './styles';
