+++ Composer/packages/client/src/pages/home/RecentBotList.tsx
@@ -17,7 +17,7 @@ import {
 } from 'office-ui-fabric-react/lib/DetailsList';
 import formatMessage from 'format-message';
 
-import { calculateTimeDiff } from '../../utils';
+import { calculateTimeDiff } from '../../utils/fileUtil';
 
 import { detailListContainer, tableCell, content } from './styles';
