+++ Composer/packages/client/src/pages/language-understanding/LUPage.tsx
@@ -1,43 +1,43 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
-
 /** @jsx jsx */
 import { jsx } from '@emotion/core';
-import React, { useContext, Fragment, useMemo, useCallback, Suspense, useEffect } from 'react';
+import React, { useContext, Fragment, useMemo, Suspense, useCallback, useEffect } from 'react';
 import formatMessage from 'format-message';
 import { Toggle } from 'office-ui-fabric-react/lib/Toggle';
 import { RouteComponentProps, Router } from '@reach/router';
 
-import { LoadingSpinner } from '../../components/LoadingSpinner/LoadingSpinner';
 import { StoreContext } from '../../store';
-import { actionButton } from '../language-understanding/styles';
-import { navigateTo } from '../../utils';
+import { navigateTo } from '../../utils/navigation';
+import { LoadingSpinner } from '../../components/LoadingSpinner/LoadingSpinner';
 import { TestController } from '../../components/TestController/TestController';
 import { INavTreeItem } from '../../components/NavTree/NavTree';
 import { Page } from '../../components/Page/Page';
 
 import TableView from './table-view';
+import { actionButton } from './styles';
 const CodeEditor = React.lazy(() => import('./code-editor'));
 
-interface LGPageProps extends RouteComponentProps<{}> {
+interface LUPageProps extends RouteComponentProps<{}> {
   dialogId?: string;
+  path: string;
 }
 
-const LGPage: React.FC<LGPageProps> = (props) => {
+const LUPage: React.FC<LUPageProps> = (props) => {
   const { state } = useContext(StoreContext);
   const { dialogs, projectId } = state;
-
   const path = props.location?.pathname ?? '';
   const { dialogId = '' } = props;
   const edit = /\/edit(\/)?$/.test(path);
+  const isRoot = dialogId === 'all';
 
   const navLinks: INavTreeItem[] = useMemo(() => {
     const newDialogLinks: INavTreeItem[] = dialogs.map((dialog) => {
       return {
         id: dialog.id,
+        url: `/bot/${projectId}/language-understanding/${dialog.id}`,
         name: dialog.displayName,
-        ariaLabel: formatMessage('language generation file'),
-        url: `/bot/${projectId}/language-generation/${dialog.id}`,
+        ariaLabel: formatMessage('language understanding file'),
       };
     });
     const mainDialogIndex = newDialogLinks.findIndex((link) => link.id === 'Main');
@@ -47,24 +47,24 @@ const LGPage: React.FC<LGPageProps> = (props) => {
       newDialogLinks.splice(0, 0, mainDialog);
     }
     newDialogLinks.splice(0, 0, {
-      id: 'common',
-      name: 'All',
-      ariaLabel: formatMessage('all language generation files'),
-      url: `/bot/${projectId}/language-generation/common`,
+      id: 'all',
+      name: formatMessage('All'),
+      ariaLabel: formatMessage('all language understanding files'),
+      url: `/bot/${projectId}/language-understanding/all`,
     });
     return newDialogLinks;
   }, [dialogs]);
 
   useEffect(() => {
     const activeDialog = dialogs.find(({ id }) => id === dialogId);
-    if (!activeDialog && dialogs.length && dialogId !== 'common') {
-      navigateTo(`/bot/${projectId}/language-generation/common`);
+    if (!activeDialog && dialogId !== 'all' && dialogs.length) {
+      navigateTo(`/bot/${projectId}/language-understanding/all`);
     }
   }, [dialogId, dialogs, projectId]);
 
   const onToggleEditMode = useCallback(
     (_e, checked) => {
-      let url = `/bot/${projectId}/language-generation/${dialogId}`;
+      let url = `/bot/${projectId}/language-understanding/${dialogId}`;
       if (checked) url += `/edit`;
       navigateTo(url);
     },
@@ -80,32 +80,36 @@ const LGPage: React.FC<LGPageProps> = (props) => {
   ];
 
   const onRenderHeaderContent = () => {
-    return (
-      <Toggle
-        checked={!!edit}
-        className={'toggleEditMode'}
-        css={actionButton}
-        defaultChecked={false}
-        offText={formatMessage('Edit mode')}
-        onChange={onToggleEditMode}
-        onText={formatMessage('Edit mode')}
-      />
-    );
+    if (!isRoot || edit) {
+      return (
+        <Toggle
+          checked={!!edit}
+          className={'toggleEditMode'}
+          css={actionButton}
+          defaultChecked={false}
+          offText={formatMessage('Edit mode')}
+          onChange={onToggleEditMode}
+          onText={formatMessage('Edit mode')}
+        />
+      );
+    }
+
+    return null;
   };
 
   return (
     <Page
-      data-testid="LGPage"
-      mainRegionName={formatMessage('LG editor')}
+      data-testid="LUPage"
+      mainRegionName={formatMessage('LU editor')}
       navLinks={navLinks}
-      navRegionName={formatMessage('LG Navigation Pane')}
-      title={formatMessage('Bot Responses')}
+      navRegionName={formatMessage('LU Navigation Pane')}
+      title={formatMessage('User Input')}
       toolbarItems={toolbarItems}
       onRenderHeaderContent={onRenderHeaderContent}
     >
       <Suspense fallback={<LoadingSpinner />}>
         <Router component={Fragment} primary={false}>
-          <CodeEditor dialogId={dialogId} path="/edit/*" />
+          <CodeEditor dialogId={dialogId} path="/edit" />
           <TableView dialogId={dialogId} path="/" />
         </Router>
       </Suspense>
@@ -113,4 +117,4 @@ const LGPage: React.FC<LGPageProps> = (props) => {
   );
 };
 
-export default LGPage;
+export default LUPage;
