+++ Composer/packages/client/src/pages/setting/SettingsPage.tsx
@@ -12,7 +12,7 @@ import { Text } from 'office-ui-fabric-react/lib/Text';
 import { StoreContext } from '../../store';
 import { TestController } from '../../components/TestController/TestController';
 import { OpenConfirmModal } from '../../components/Modal/ConfirmDialog';
-import { navigateTo } from '../../utils';
+import { navigateTo } from '../../utils/navigation';
 import { Page } from '../../components/Page/Page';
 import { INavTreeItem } from '../../components/NavTree/NavTree';
 import { useLocation } from '../../utils/hooks';
