+++ Composer/packages/client/src/router.tsx
@@ -13,12 +13,12 @@ import { BASEPATH } from './constants';
 import { StoreContext } from './store';
 import { LoadingSpinner } from './components/LoadingSpinner/LoadingSpinner';
 
-const DesignPage = React.lazy(() => import('./pages/design'));
-const LUPage = React.lazy(() => import('./pages/language-understanding'));
-const LGPage = React.lazy(() => import('./pages/language-generation'));
-const SettingPage = React.lazy(() => import('./pages/setting'));
-const Notifications = React.lazy(() => import('./pages/notifications'));
-const Publish = React.lazy(() => import('./pages/publish'));
+const DesignPage = React.lazy(() => import('./pages/design/DesignPage'));
+const LUPage = React.lazy(() => import('./pages/language-understanding/LUPage'));
+const LGPage = React.lazy(() => import('./pages/language-generation/LGPage'));
+const SettingPage = React.lazy(() => import('./pages/setting/SettingsPage'));
+const Notifications = React.lazy(() => import('./pages/notifications/Notifications'));
+const Publish = React.lazy(() => import('./pages/publish/Publish'));
 const Skills = React.lazy(() => import('./pages/skills'));
 const BotCreationFlowRouter = React.lazy(() => import('./components/CreationFlow/CreationFlow'));
