--- Composer/packages/client/src/utils/index.ts
@@ -1,6 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-export * from './fileUtil';
-export * from './dialogUtil';
-export * from './navigation';
