+++ Composer/packages/client/src/utils/navigation.ts
@@ -7,7 +7,9 @@ import { navigate, NavigateOptions } from '@reach/router';
 import { BreadcrumbItem, DesignPageLocation } from '../store/types';
 import { BASEPATH } from '../constants';
 
-import { parsePathToFocused, parsePathToSelected, parseTypeToFragment } from './convertUtils';
+import { parsePathToFocused } from './convertUtils/parsePathToFocused';
+import { parsePathToSelected } from './convertUtils/parsePathToSelected';
+import { parseTypeToFragment } from './convertUtils/parseTypeToFragment';
 import { resolveToBasePath } from './fileUtil';
 export const BreadcrumbUpdateType = {
   Selected: 'selected',
